:wrench: re-organizes the sources and gives some modules clearer (less) responsisbilities
