+++ .eslintignore
@@ -4,4 +4,5 @@ tmp
 utl
 src/report/*.template.js
 test/*/fixtures/**
+test/*/*/fixtures/**
 src/report/lib/**
