+++ src/cli/rules.starter.json
