+++ src/extract/extract.js
@@ -5,9 +5,9 @@ const acorn_loose                 = require('acorn/dist/acorn_loose');
 const fs                          = require('fs');
 const _                           = require('lodash');
 const path                        = require('path');
-const resolve                     = require('../resolve');
+const resolve                     = require('./resolve');
 const validate                    = require('../validate');
-const transpile                   = require('../transpile');
+const transpile                   = require('./transpile');
 const utl                         = require('../utl');
 const extractES6Dependencies      = require('./extract-ES6');
 const extractCommonJSDependencies = require('./extract-commonJS');
