+++ src/extract/resolve/determineDependencyTypes.js
