+++ src/extract/resolve/index.js
