+++ src/extract/resolve/resolve-commonJS.js
