+++ src/extract/transpile/javaScriptWrap.js
