+++ src/extract/transpile/tryRequire.js
