+++ src/extract/transpile/typeScriptWrap.js
@@ -2,7 +2,7 @@
 const tryRequire = require("./tryRequire");
 const typescript = tryRequire(
     "typescript",
-    require("../../package.json").supportedTranspilers.typescript
+    require("../../../package.json").supportedTranspilers.typescript
 );
 
 exports.isAvailable = () => typescript !== false;
