+++ test/extract/resolve/fixtures/empty-package-json/package.json
