+++ test/extract/resolve/fixtures/invalid-package-json/package.json
