+++ test/extract/resolve/fixtures/minimal-package-json/package.json
