+++ test/extract/transpile/fixtures/coffee.coffee
