+++ test/extract/transpile/fixtures/coffee.js
