+++ test/extract/transpile/fixtures/litcoffee.js
