+++ test/extract/transpile/fixtures/markdownlitcoffee.js
