+++ test/extract/transpile/fixtures/typescriptscript.js
