+++ test/extract/transpile/fixtures/typescriptscript.ts
