+++ test/extract/transpile/liveScriptWrap.spec.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const expect = require("chai").expect;
-const wrap   = require("../../src/transpile/liveScriptWrap");
+const wrap   = require("../../../src/extract/transpile/liveScriptWrap");
 
 describe("livescript transpiler", () => {
     it("tells the livescript transpiler is not available", () => {
