+++ test/extract/transpile/typeScriptWrap.spec.js
@@ -2,7 +2,7 @@
 
 const expect = require("chai").expect;
 const fs     = require('fs');
-const wrap  = require("../../src/transpile/typeScriptWrap");
+const wrap  = require("../../../src/extract/transpile/typeScriptWrap");
 
 describe("typescript transpiler", () => {
     it("tells the typescript transpiler is available", () => {
@@ -14,10 +14,10 @@ describe("typescript transpiler", () => {
     it("transpiles typescript", () => {
         expect(
             wrap.transpile(
-                fs.readFileSync("./test/transpile/fixtures/typescriptscript.ts", 'utf8')
+                fs.readFileSync("./test/extract/transpile/fixtures/typescriptscript.ts", 'utf8')
             )
         ).to.equal(
-            fs.readFileSync("./test/transpile/fixtures/typescriptscript.js", 'utf8')
+            fs.readFileSync("./test/extract/transpile/fixtures/typescriptscript.js", 'utf8')
         );
     });
 });
